app/requirements.txt
flask
mysql-connector-python
requests
pytest